+++ frontend/components/user/UserProfile.tsx
@@ -7,8 +7,16 @@ const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
+  width: 100%;
+
+  @media (max-width: 960px) {
+    flex-direction: row;
+    /* justify-content: space-between; */
+    gap: 20px;
+  }
 `;
 export const AvatarContainer = styled.div`
+  flex-shrink: 0;
   display: flex;
   justify-content: center;
   width: 150px;
@@ -19,6 +27,33 @@ export const AvatarContainer = styled.div`
   box-shadow: 5px 5px 10px 0px rgba(0, 0, 0, 0.25);
   -webkit-box-shadow: 5px 5px 10px 0px rgba(0, 0, 0, 0.25);
   -moz-box-shadow: 5px 5px 10px 0px rgba(0, 0, 0, 0.25);
+`;
+const ProfileContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  align-items: center;
+  gap: 10px;
+  @media (max-width: 960px) {
+    /* flex-direction: row; */
+    align-items: flex-start;
+  }
+`;
+const Position = styled.div.attrs({
+  className: 'noto-regular-13',
+})`
+  display: flex;
+  flex-direction: column;
+  p {
+    text-align: center;
+  }
+
+  @media (max-width: 960px) {
+    align-items: center;
+    flex-direction: row;
+    gap: 10px;
+  }
 `;
 const Name = styled.p.attrs({
   className: 'nanum-bold',
@@ -32,7 +67,7 @@ const StackWrapper = styled.div.attrs({})`
   width: 100%;
   gap: 8px;
   margin-top: 20px;
-  margin-bottom: 50px;
+  margin-bottom: 20px;
   color: white;
 `;
 const stacks = [
@@ -56,15 +91,18 @@ export default function UserProfile({ user }: { user: UserState }) {
           />
         )}
       </AvatarContainer>
-      <Tag>{`${user?.yearOfDev} 년차`}</Tag>
-      <Name>{user?.name}</Name>
-      <p className="noto-regular-13">프론트엔드</p>
-      <StackWrapper>
-        {stacks.map((stack) => (
-          // <Tag key={`${stack}`}>{stack}</Tag>
-          <Stack key={stack} tech={stack} />
-        ))}
-      </StackWrapper>
+      <ProfileContainer>
+        <Name>{user?.name}</Name>
+        <Position>
+          <p>{`${user?.yearOfDev} 년차`}</p>
+          <p className="noto-regular-13">프론트엔드</p>
+        </Position>
+        <StackWrapper>
+          {stacks.map((stack) => (
+            <Stack key={stack} tech={stack} />
+          ))}
+        </StackWrapper>
+      </ProfileContainer>
     </Wrapper>
   );
 }
